fix: clean up frontend code issues
- Add useClickOutside hook to deduplicate click-outside logic
- Move formatDuration to shared utils, remove duplicate implementations
- Fix dual error notifications: use inline for validation, toast for API errors
- Remove unused state setter hack for Spotify restore (use query invalidation)
- Add refetchIntervalInBackground: false to stop polling when tab hidden
- Remove unused API method (getProhibitedPatterns) and auth method (setAuth)
